prep for 2.0.0
git-svn-id: svn+ssh://c99.org/home/c99org/svn/DCSquares/trunk@51 8944af1b-4df6-0310-90e0-d6e02aab4a67

+++ common/http.cc
@@ -117,7 +117,7 @@ uint32_t resolve(char *name) {
 }
 #endif
 #ifdef UNIX
-uint32_t resolve(char *name) {
+long unsigned int resolve(char *name) {
   char tmp[200];
   struct hostent *h=gethostbyname(name);
   if(h==NULL) {
